requirements-docker.txt
# Image Linux : pas de python-magic-bin (Windows uniquement)

langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0

lancedb>=0.10.0

sentence-transformers==3.0.0
huggingface-hub==0.21.0

ollama==0.2.0

pypdf==4.0.0
unstructured==0.11.8
python-magic==0.4.27

fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.8

python-dotenv==1.0.0
tqdm==4.67.0
